Parse command line parameters and login

+++ cgfs.c
@@ -1,3 +1,4 @@
+#include "cgapi.h"
 #include "dict.h"
 
 #include <errno.h>
@@ -8,6 +9,7 @@
 #include <string.h>
 
 #include <sys/types.h>
+#include <termios.h>
 #include <time.h>
 #include <unistd.h>
 
@@ -45,6 +47,13 @@
 
 #define UNUSED(x) (void)(x);
 
+struct cgfs_context {
+        cgapi_token_t login_token;
+        struct dict *open_files;
+};
+
+#define CGFS_CONTEXT ((struct cgfs_context *) fuse_get_context()->private_data)
+
 struct file {
         int id;
         size_t nlinks;
@@ -60,8 +69,6 @@ struct file {
         char *buf;
 };
 
-struct dict open_files;
-
 int cgfs_getattr(const char *path, struct stat *st)
 {
         UNUSED(path);
@@ -74,7 +81,7 @@ int cgfs_getattr(const char *path, struct stat *st)
 
         // TODO: Get other info from server
 
-        struct file *f = dict_get(&open_files, path);
+        struct file *f = dict_get(CGFS_CONTEXT->open_files, path);
         if (f) {
                 // st->st_mtime = f->mtime;
         } else {
@@ -86,7 +93,7 @@ int cgfs_getattr(const char *path, struct stat *st)
 
 int cgfs_access(const char *path, int mask)
 {
-        struct file *f = dict_get(&open_files, path);
+        struct file *f = dict_get(CGFS_CONTEXT->open_files, path);
         if (!f) {
                 // Get file from server
                 if (!f) {
@@ -170,12 +177,12 @@ int cgfs_open(const char *path, struct fuse_file_info *fi)
 {
         UNUSED(fi);
 
-        struct file *f = dict_get(&open_files, path);
+        struct file *f = dict_get(CGFS_CONTEXT->open_files, path);
 
         if (!f) {
                 MALLOC(struct file, f, 1);
                 // TODO: Get data from server
-                dict_set(&open_files, path, f);
+                dict_set(CGFS_CONTEXT->open_files, path, f);
         }
 
         f->nlinks++;
@@ -187,12 +194,12 @@ int cgfs_release(const char *path, struct fuse_file_info *fi)
 {
         UNUSED(fi);
 
-        struct file *f = dict_get(&open_files, path);
+        struct file *f = dict_get(CGFS_CONTEXT->open_files, path);
         if (!f) return EBADF;
 
         f->nlinks--;
         if (f->nlinks == 0) {
-                dict_unset(&open_files, path);
+                dict_unset(CGFS_CONTEXT->open_files, path);
                 FREE(f);
         }
 
@@ -204,7 +211,7 @@ int cgfs_read(const char *path, char *buf, size_t buflen, off_t offset,
 {
         UNUSED(fi);
 
-        struct file *f = dict_get(&open_files, path);
+        struct file *f = dict_get(CGFS_CONTEXT->open_files, path);
         if (!f) return EBADF;
 
         if ((size_t)offset > f->buflen) {
@@ -225,7 +232,7 @@ int cgfs_write(const char *path, const char *buf, size_t buflen, off_t offset,
 {
         UNUSED(fi);
 
-        struct file *f = dict_get(&open_files, path);
+        struct file *f = dict_get(CGFS_CONTEXT->open_files, path);
         if (!f) return EBADF;
 
         REALLOC(char, f->buf, f->buflen + buflen);
@@ -240,7 +247,7 @@ int cgfs_ftruncate(const char *path, off_t offset, struct fuse_file_info *fi)
 {
         UNUSED(fi);
 
-        struct file *f = dict_get(&open_files, path);
+        struct file *f = dict_get(CGFS_CONTEXT->open_files, path);
         if (!f) return EBADF;
 
         REALLOC(char, f->buf, offset);
@@ -252,7 +259,7 @@ int cgfs_flush(const char *path, struct fuse_file_info *fi)
 {
         UNUSED(fi);
 
-        struct file *f = dict_get(&open_files, path);
+        struct file *f = dict_get(CGFS_CONTEXT->open_files, path);
         if (!f) return EBADF;
 
         if (!f->dirty) {
@@ -275,24 +282,77 @@ int cgfs_unlink(const char *path)
         return 0;
 }
 
-static struct fuse_operations fuse_ops = {
-        .getattr = cgfs_getattr,
-        .access = cgfs_access,
-        .mkdir = cgfs_mkdir,
-        .rmdir = cgfs_rmdir,
-        .readdir = cgfs_readdir,
-        .open = cgfs_open,
-        .release = cgfs_release,
-        .read = cgfs_read,
-        .write = cgfs_write,
-        .ftruncate = cgfs_ftruncate,
-        .flush = cgfs_flush,
-        .unlink = cgfs_unlink,
-};
+// Read a nul-terminated password into the buffer
+void get_password(size_t len, char pass[len])
+{
+        printf("Enter password: ");
+
+	struct termios oflags, nflags;
+	tcgetattr(fileno(stdin), &oflags);
+	nflags = oflags;
+	nflags.c_lflag &= ~ECHO;
+	nflags.c_lflag |= ECHONL;
+
+	if (tcsetattr(fileno(stdin), TCSANOW, &nflags) != 0) {
+		perror("tcsetattr");
+		exit(EXIT_FAILURE);
+	}
+
+	size_t i = 0;
+	int c;
+	while ((c = getchar()) != EOF && c != '\n' && i < len - 1)
+		pass[i++] = c;
+
+	pass[i] = '\0';
+
+	if (tcsetattr(fileno(stdin), TCSANOW, &oflags) != 0) {
+		perror("tcsetattr");
+		exit(EXIT_FAILURE);
+	}
+}
 
 int main(int argc, char *argv[argc])
 {
-        dict_init(&open_files, 1 << 8);
+        if (argc < 3 || argv[argc - 2][0] == '-' || argv[argc - 1][0] == '-') {
+                fprintf(stderr, "%s: last 2 arguments must be mountpoint and email, respectively\n", argv[0]);
+                exit(EXIT_FAILURE);
+        } 
 
-        return fuse_main(argc, argv, &fuse_ops, NULL);
+        char *email = argv[--argc];
+        argv[argc] = NULL;
+
+        char password[128];
+        get_password(128, password);
+
+        static struct cgfs_context context;
+
+        context.login_token = cgapi_login(email, password);
+        if (context.login_token == NULL) {
+                fprintf(stderr, "%s: login failed\n", argv[0]);
+                exit(EXIT_FAILURE);
+        }
+
+        context.open_files = dict_create(1 << 8);
+        if (context.open_files == NULL) {
+                printf("%s: ", argv[0]);
+                perror("dict_create");
+                exit(EXIT_FAILURE);
+        }
+
+        static struct fuse_operations fuse_ops = {
+                .getattr = cgfs_getattr,
+                .access = cgfs_access,
+                .mkdir = cgfs_mkdir,
+                .rmdir = cgfs_rmdir,
+                .readdir = cgfs_readdir,
+                .open = cgfs_open,
+                .release = cgfs_release,
+                .read = cgfs_read,
+                .write = cgfs_write,
+                .ftruncate = cgfs_ftruncate,
+                .flush = cgfs_flush,
+                .unlink = cgfs_unlink,
+        };
+
+        return fuse_main(argc, argv, &fuse_ops, &context);
 }
